Run the xmpp stuff in the background
This commit uses hook_process to run the xmpp call
in the background, so it doesn't block weechat.

Also, this commit updates the jid password to use
/secure functionality.

Finally, this commit adds a new command /hl that
toggles on/off the highlightxmpp functionality.
If you are in a PM session with another user, you
probably want to turn it off.